+++ CoolProp/gitrevision.h
@@ -1,1 +1,1 @@
-std::string gitrevision = "9a179be2c74c667c262dd578b4f43ec569398cb4";
+std::string gitrevision = "b'2c7ac1777e9c757b666ef69d3176e28a877219af'";